+++ src/main/java/database/StubTipDao.java
