+++ src/main/java/tiplogic/Course.java
