+++ src/main/java/tiplogic/Mappable.java
