+++ src/main/java/tiplogic/Tip.java
