+++ src/main/java/tiplogic/Vinkki.java
@@ -1,4 +1,4 @@
-package vinkkilogic;
+package tiplogic;
 
 import java.util.List;
