+++ src/test/java/tiplogic/KirjaTest.java
